netjack1/2 Opus: use only 2bytes for encoded-length

+++ common/JackNetTool.h
@@ -419,7 +419,7 @@ namespace Jack
             OpusCustomEncoder** fOpusEncoder;
             OpusCustomDecoder** fOpusDecoder;
 
-            size_t *fCompressedSizesByte;
+            unsigned short *fCompressedSizesByte;
             int fCompressedMaxSizeByte;
             int fNumPackets;
